refactored regressor into base and inherited class and added classifier inheriting from same base
